comment

+++ src/main/resources/db/migration/V1__Create_user_table.sql
@@ -1,6 +1,6 @@
 create table USER
 (
-	ID INT auto_increment primary key not null,
+	ID BIGINT auto_increment primary key not null,
 	ACCOUNT_ID VARCHAR(100),
 	NAME VARCHAR(50),
 	TOKEN CHAR(36),
