+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -1,11 +1,11 @@
 create table question
 (
-	id int auto_increment primary key,
+	id BIGINT auto_increment primary key,
 	title varchar(50),
 	description text,
 	gmt_create bigint,
-	gmt_modify int,
-	creator int,
+	gmt_modify bigint,
+	creator bigint,
 	comment_count int default 0,
 	view_count int default 0,
 	like_count int default 0,
